+++ pdata/plog/generated_logs_test.go
@@ -486,8 +486,10 @@ func TestLogRecord_SpanID(t *testing.T) {
 
 func TestLogRecord_FlagsStruct(t *testing.T) {
 	ms := NewLogRecord()
-	internal.FillTestLogRecordFlags(internal.LogRecordFlags(ms.FlagsStruct()))
-	assert.Equal(t, LogRecordFlags(internal.GenerateTestLogRecordFlags()), ms.FlagsStruct())
+	assert.Equal(t, LogRecordFlags(0), ms.FlagsStruct())
+	testValFlagsStruct := LogRecordFlags(1)
+	ms.SetFlagsStruct(testValFlagsStruct)
+	assert.Equal(t, testValFlagsStruct, ms.FlagsStruct())
 }
 
 func TestLogRecord_SeverityText(t *testing.T) {